Add utility function for fetching data from Strapi
Note that this does not set up anything like server actions or SWR. I'm not sure how the repo will be structured so I didn't want to make assumptions.

Also, note that this is not an opinionated setup. I think some opinions could be useful. For example, we may _always_ want to request Events with the same `?populate` query param. Some infrastructure could be created for that.